+++ trunk/game/sprites/SpriteViewer.cs
@@ -33,7 +33,7 @@ namespace AbrahmanAdventure.sprites
         /// <param name="isOddFrame">whether frame is odd</param>
         /// <param name="playerSprite">player's sprite</param>
         /// <param name="visibleSpriteList">list of currently visible sprites</param>
-        internal void Update(double viewOffsetX, double viewOffsetY, AbstractSprite playerSprite, HashSet<AbstractSprite> visibleSpriteList, bool isOddFrame)
+        internal void Update(float viewOffsetX, float viewOffsetY, AbstractSprite playerSprite, HashSet<AbstractSprite> visibleSpriteList, bool isOddFrame)
         {
             foreach (AbstractSprite sprite in visibleSpriteList)
                 if (sprite != playerSprite)
@@ -51,17 +51,17 @@ namespace AbrahmanAdventure.sprites
         /// <param name="viewOffsetX">view offset X</param>
         /// <param name="viewOffsetY">view offset Y</param>
         /// <param name="isOddFrame">whether frame is odd</param>
-        private void ShowSprite(AbstractSprite sprite, double viewOffsetX, double viewOffsetY, bool isOddFrame)
+        private void ShowSprite(AbstractSprite sprite, float viewOffsetX, float viewOffsetY, bool isOddFrame)
         {
             if (isOddFrame && (sprite.HitCycle.IsFired || sprite is PlayerSprite && ((PlayerSprite)sprite).FromVortexCycle.IsFired))
                 return;
 
-            double specialOffsetX, specialOffsetY;
+            float specialOffsetX, specialOffsetY;
 
             Surface spriteSurface = sprite.GetCurrentSurface(out specialOffsetX, out specialOffsetY);
 
-            int xBlitPosition = (int)Math.Round(((sprite.XPosition - ((double)spriteSurface.GetWidth() / (double)Program.tileSize) / 2.0 - viewOffsetX + specialOffsetX) * Program.tileSize));
-            int yBlitPosition = (int)((sprite.YPosition - viewOffsetY + specialOffsetY) * (double)Program.tileSize) - spriteSurface.GetHeight();
+            int xBlitPosition = (int)Math.Round(((sprite.XPosition - ((float)spriteSurface.GetWidth() / (float)Program.tileSize) / 2.0f - viewOffsetX + specialOffsetX) * Program.tileSize));
+            int yBlitPosition = (int)((sprite.YPosition - viewOffsetY + specialOffsetY) * (float)Program.tileSize) - spriteSurface.GetHeight();
 
             mainSurface.Blit(spriteSurface, new Point(xBlitPosition, yBlitPosition),spriteSurface.GetRectangle());
         }
